fixup! refactor(core): improve code around element validation
resolve circular dependencies issue
